requirements.txt
requests>=2.25.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
webdriver-manager==4.0.1
streamlit==1.28.1
openpyxl==3.1.2
plotly>=5.0.0
xlrd==2.0.1
pymongo==4.6.1
python-dotenv==1.0.0
supabase>=2.7.0
gotrue>=2.0.0